First base of animated Cuphead
Just made a quick spritesheet cuphead in the main to test how everything works.
Next on the agenda, make it in his own class and add gravity, input, collisions.

(I'm on Erasmus so it's difficult to make time, but I hope to finish it at the end of this week!)